# Prefetch hash entries