--- a/src/transposition.h
+++ b/src/transposition.h
@@ -75,6 +75,7 @@ void clearTT();
 int hashfullTT();
 int valueFromTT(int value, int height);
 int valueToTT(int value, int height);
+void prefetchTTEntry(uint64_t hash);
 int getTTEntry(uint64_t hash, uint16_t *move, int *value, int *eval, int *depth, int *bound);
 void storeTTEntry(uint64_t hash, uint16_t move, int value, int eval, int depth, int bound);