<pre>ELO   | 6.27 +- 4.41 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=8MB
LLR   | 3.02 (-2.94, 2.94) [0.00, 5.00]
Games | N: 10027 W: 2202 L: 2021 D: 5804
http://chess.grantnet.us/viewTest/4578/

ELO   | 4.24 +- 3.26 (95%)
SPRT  | 5.0+0.05s Threads=8 Hash=64MB
LLR   | 3.01 (-2.94, 2.94) [0.00, 5.00]
Games | N: 16700 W: 3308 L: 3104 D: 10288
http://chess.grantnet.us/viewTest/4580/

NO FUNCTIONAL CHANGE

BENCH : 7,681,536</pre>

The only diff between the commit used in testing is the VERSION_ID bump in uci.h.

Thanks to @TerjeKir for rebasing my original prefetch patch and fixing a small whitespace issue, and @Alayan-stk-2 for further pointers on preparing the PR.