@skiminki 2020-02-09
I ran some quick benchmarks on my box (i7-8870k) with the following commands: (3 runs each)

<pre>  setoption name Threads value 1
  setoption name Hash value <Hash>
  ucinewgame
  isready
  go movetime 10000</pre>

<pre>Hash | 11.91 nps      | prefetch nps    | speedup
-----+----------------+-----------------+--------
128  | 2609333        | 2661000         | +1.98%
4096 | 2578000        | 2629000         | +1.90%</pre>

I'm actually a bit surprised that the test result is a good as it is. Perhaps prefetch is more useful with the HW in openbench.