@Alayan-stk-2 2020-02-09
You need 35 2% speedups to get a 100% speedup. At 8s+0.08, a 100% speedup is worth something like 150 elo. 150/35 = ~4.3.

So, that's rather expected. Of course, the elo gain will be less impressive at longer TCs, but it's still a nice boost, and it also speeds up future patch testing.